Restructure Serial/ThreeSolvers
