--- src/precice/tests/mapping-nearest-projection.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<precice-configuration>
-  <solver-interface dimensions="3">
-    <data:scalar name="DataOne" />
-
-    <mesh name="MeshOne">
-      <use-data name="DataOne" />
-    </mesh>
-
-    <mesh name="MeshTwo">
-      <use-data name="DataOne" />
-    </mesh>
-
-    <participant name="SolverOne">
-      <use-mesh name="MeshOne" provide="on" />
-      <write-data name="DataOne" mesh="MeshOne" />
-    </participant>
-
-    <participant name="SolverTwo">
-      <use-mesh name="MeshOne" from="SolverOne" />
-      <use-mesh name="MeshTwo" provide="on" />
-      <mapping:nearest-projection
-        direction="read"
-        from="MeshOne"
-        to="MeshTwo"
-        constraint="consistent"
-        timing="initial" />
-      <read-data name="DataOne" mesh="MeshTwo" />
-    </participant>
-
-    <m2n:sockets from="SolverOne" to="SolverTwo" />
-
-    <coupling-scheme:serial-explicit>
-      <participants first="SolverOne" second="SolverTwo" />
-      <max-time-windows value="1" />
-      <time-window-size value="1.0" />
-      <exchange data="DataOne" mesh="MeshOne" from="SolverOne" to="SolverTwo" />
-    </coupling-scheme:serial-explicit>
-  </solver-interface>
-</precice-configuration>
