+++ tests/serial/three-solvers/ThreeSolversImplicitImplicit.cpp
@@ -1,0 +1,23 @@
+#ifndef PRECICE_NO_MPI
+
+#include "testing/Testing.hpp"
+
+#include <precice/SolverInterface.hpp>
+#include "helpers.hpp"
+
+BOOST_AUTO_TEST_SUITE(PreciceTests)
+BOOST_AUTO_TEST_SUITE(Serial)
+BOOST_AUTO_TEST_SUITE(ThreeSolvers)
+BOOST_AUTO_TEST_CASE(ThreeSolversImplicitImplicit)
+{
+  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank), "SolverThree"_on(1_rank));
+  std::string      config = context.config();
+  std::vector<int> expectedCallsOfAdvance{30, 30, 20};
+  runTestThreeSolvers(config, expectedCallsOfAdvance, context);
+}
+
+BOOST_AUTO_TEST_SUITE_END() // PreciceTests
+BOOST_AUTO_TEST_SUITE_END() // Serial
+BOOST_AUTO_TEST_SUITE_END() // ThreeSolvers
+
+#endif // PRECICE_NO_MPI
