+++ tests/serial/three-solvers/ThreeSolversParallel.xml
