+++ tests/tests.cmake
@@ -83,6 +83,13 @@ target_sources(testprecice
     tests/serial/multi-coupling/MultiCouplingThreeSolvers3.cpp
     tests/serial/multi-coupling/helpers.cpp
     tests/serial/multi-coupling/helpers.hpp
+    tests/serial/three-solvers/ThreeSolversExplicitExplicit.cpp
+    tests/serial/three-solvers/ThreeSolversExplicitImplicit.cpp
+    tests/serial/three-solvers/ThreeSolversImplicitExplicit.cpp
+    tests/serial/three-solvers/ThreeSolversImplicitImplicit.cpp
+    tests/serial/three-solvers/ThreeSolversParallel.cpp
+    tests/serial/three-solvers/helpers.cpp
+    tests/serial/three-solvers/helpers.hpp
     tests/serial/time/explicit/DoNothingWithSubcycling.cpp
     tests/serial/time/explicit/ReadWriteScalarDataWithSubcycling.cpp
     tests/serial/time/implicit/ReadWriteScalarDataWithSubcycling.cpp
